requirements.txt
ipython
pyyaml
sensei_client
pyserial
redis
